Add files via upload
"Notencrypt" is a comprehensive and secure note-taking computer application that prioritizes data security and privacy while providing advanced features for convenience and organization. The problem statement is the need for secure communication and data storage. With the increase in data breaches, protecting personal data has become a top priority for individuals and organizations alike. Notencrypt offers a solution to this problem by providing a platform for encrypted communication and data storage.